requirements.txt
alpaca.py
pipenv